# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Skolara API",
    "description": "Backend for searching colleges, managing applications, and processing payments.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:8000/api" },
    { "url": "https://devapi.skolara.in/api" }
  ],
...
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/payments/create-order": {
      "post": {
        "tags": ["Payments"],
        "summary": "Create a Razorpay order for an application",
        "operationId": "createOrder",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateOrderRequest" },
              "examples": {
                "default": {
                  "value": {
                    "application_id": "5c3ed8b2-9a0e-4c2e-8f81-8b0b1e8e0d77"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateOrderResponse"
                },
                "examples": {
                  "ok": {
                    "value": {
                      "success": true,
                      "message": "Order created successfully",
                      "order": {
                        "id": "ord_123",
                        "provider_order_id": "order_RP_ABC123",
                        "amount": 500000,
                        "currency": "INR",
                        "status": "Created"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or application not payable",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/JWTUnauthorized" },
          "404": {
            "description": "Application not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": {
            "description": "Order creation failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/payments/verify-payment": {
      "post": {
        "tags": ["Payments"],
        "summary": "Verify a captured Razorpay payment and mark order/application paid",
        "operationId": "verifyPayment",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/VerifyPaymentRequest" },
              "examples": {
                "default": {
                  "value": {
                    "razorpay_payment_id": "pay_29QQoUBi66xm2f",
                    "razorpay_order_id": "order_DBJOWzybf0sJbb",
                    "razorpay_signature": "generated_signature_here"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment verified",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccessResponse"
                },
                "examples": {
                  "ok": {
                    "value": { "success": true, "message": "Payment verified" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid payload / signature mismatch / not captured / amount mismatch",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/JWTUnauthorized" },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": {
            "description": "Verification failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/payments/update-payment-failure": {
      "post": {
        "tags": ["Payments"],
        "summary": "Record a failed payment attempt for an order",
        "operationId": "updatePaymentFailure",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePaymentFailureRequest"
              },
              "examples": {
                "default": {
                  "value": {
                    "order_id": "ord_123",
                    "provider_order_id": "order_DBJOWzybf0sJbb",
                    "provider_payment_id": "pay_29QQoUBi66xm2f",
                    "reason": "Payment failed due to insufficient funds"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Failure recorded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccessResponse"
                },
                "examples": {
                  "ok": {
                    "value": {
                      "success": true,
                      "message": "Payment failure recorded"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid order details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/JWTUnauthorized" },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": {
            "description": "Recording failure failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/payments/cancel-order": {
      "post": {
        "tags": ["Payments"],
        "summary": "Cancel an order if not completed",
        "operationId": "cancelOrder",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CancelOrderRequest" },
              "examples": {
                "default": {
                  "value": {
                    "order_id": "ord_123",
                    "provider_order_id": "order_DBJOWzybf0sJbb"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order canceled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccessResponse"
                },
                "examples": {
                  "ok": {
                    "value": {
                      "success": true,
                      "message": "Order canceled successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid order details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/JWTUnauthorized" },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": {
            "description": "Cancellation failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Standard JWT bearer token (used on protected endpoints)."
      }
    },
    "responses": {
      "JWTUnauthorized": {
        "description": "Missing, malformed, expired, or invalid JWT",
        "headers": {
          "WWW-Authenticate": {
            "schema": { "type": "string" },
            "description": "Auth challenge header (e.g., Bearer realm and error details)"
          }
        },
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/JWTErrorResponse" },
            "examples": {
              "missingHeader": {
                "value": {
                  "error": "UNAUTHORIZED",
                  "message": "Missing or malformed Authorization header",
                  "statusCode": 401
                }
              },
              "invalid": {
                "value": {
                  "error": "INVALID_TOKEN",
                  "message": "JWT signature verification failed",
                  "statusCode": 401
                }
              },
              "expired": {
                "value": {
                  "error": "TOKEN_EXPIRED",
                  "message": "JWT access token has expired",
                  "statusCode": 401
                }
              }
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        },
        "required": ["error"]
      },
...
      "JWTErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "enum": [
              "UNAUTHORIZED",
              "INVALID_TOKEN",
              "TOKEN_EXPIRED",
              "FORBIDDEN"
            ],
            "description": "JWT validation error type"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error details"
          },
          "statusCode": {
            "type": "integer",
            "example": 401,
            "description": "HTTP status code"
          }
        },
        "required": ["error"]
      },
...
      "CreateOrderRequest": {
        "type": "object",
        "required": ["application_id"],
        "properties": {
          "application_id": {
            "type": "string",
            "format": "uuid",
            "description": "applied_colleges.ref_id"
          }
        },
        "additionalProperties": false
      },
      "CreateOrderResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "message": { "type": "string" },
          "order": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "description": "Internal order ID (UUID)"
              },
              "provider_order_id": {
                "type": "string",
                "description": "Razorpay order id"
              },
              "amount": {
                "type": "integer",
                "description": "Amount in paisa (Razorpay minor units)",
                "example": 500000
              },
              "currency": { "type": "string", "example": "INR" },
              "status": { "type": "string", "example": "Created" }
            },
            "required": [
              "id",
              "provider_order_id",
              "amount",
              "currency",
              "status"
            ]
          }
        },
        "required": ["success", "message", "order"]
      },
      "VerifyPaymentRequest": {
        "type": "object",
        "required": [
          "razorpay_payment_id",
          "razorpay_order_id",
          "razorpay_signature"
        ],
        "properties": {
          "razorpay_payment_id": { "type": "string" },
          "razorpay_order_id": {
            "type": "string",
            "description": "Must match orders.provider_order_id"
          },
          "razorpay_signature": { "type": "string" }
        },
        "additionalProperties": false
      },
      "UpdatePaymentFailureRequest": {
        "type": "object",
        "required": ["order_id", "provider_order_id", "provider_payment_id"],
        "properties": {
          "order_id": { "type": "string", "description": "orders.id (UUID)" },
          "provider_order_id": {
            "type": "string",
            "description": "Razorpay order id"
          },
          "provider_payment_id": {
            "type": "string",
            "description": "Razorpay payment id"
          },
          "reason": { "type": "string", "nullable": true }
        },
        "additionalProperties": false
      },
      "CancelOrderRequest": {
        "type": "object",
        "required": ["order_id", "provider_order_id"],
        "properties": {
          "order_id": { "type": "string", "description": "orders.id (UUID)" },
          "provider_order_id": {
            "type": "string",
            "description": "Razorpay order id"
          }
        },
        "additionalProperties": false
      },
      "GenericSuccessResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "message": { "type": "string" }
        },
        "required": ["success", "message"]
      }
    }
  }
}